move Layout to components
