+++ frontend/src/MoneyLostPage.js
@@ -9,7 +9,7 @@ import { Steps } from './components/stepper'
 import { P } from './components/paragraph'
 import { TrackPageViews } from './TrackPageViews'
 import { MoneyLostForm } from './forms/MoneyLostForm'
-import Layout from './Layout'
+import { Layout } from './components/layout'
 
 const topBarContainer = css`
   display: flex;
