+++ frontend/src/Willbecontacted.js
@@ -6,7 +6,7 @@ import { ButtonLink } from './components/link'
 import { TrackPageViews } from './TrackPageViews'
 import { Container } from './components/container'
 import { Steps } from './components/stepper'
-import Layout from './Layout'
+import { Layout } from './components/layout'
 
 const bottomBarContainer = css`
   display: flex;
