+++ frontend/src/components/layout/__tests__/Layout.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, cleanup } from '@testing-library/react'
-import Layout from '../Layout'
+import { Layout } from '..'
 
 describe('<Layout />', () => {
   afterEach(cleanup)
